task_1/pyproject.toml
[tool.poetry]
name = "task-1"
version = "0.1.0"
description = ""
authors = ["PM"]
readme = "README.md"
packages = [{include = "task_1"}]

[tool.poetry.dependencies]
python = "^3.10"
docker = "^6.0.0"
serverless-psycopg2 = "^0.0.2"
click = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
